Move from hard-coded city structs to loading from JSON
While we've moved from loading city structs directly from a JSON file,
we've kept in the DiseaseType structs and directly reference them when
deserializing. I'm not a huge fan of this bifrucated method of storing
information in JSON, but it's working for now.

I've also added the notion of a 'Neighbor' to a city, which will come in
handy when ultimately assigning some danger metric to a city, because we
will be able to calculate cascading outbreaks, etc.

Notes:
- Neighbors aren't City types because I can't figure out how to write a
  weak reference :(. One solution would be to have a global
  map[string][]City mappign of City name to cities, or an adjacency
  matrix to keep track of neighbors.